seperate class from initialisation
